fix: исправил баг с ссылками в портфолио
