+++ src/components/App/App.css
@@ -1,3 +1,7 @@
+html {
+  scroll-behavior: smooth;
+}
+
 .page {
   background-color: #202020;
   color: #fff;
