+++ src/index.css
@@ -1,10 +1,10 @@
 @import url(./vendor/normalize.css);
-@import url(./vendor/global.css);
 @import url(./vendor/fonts/font.css);
 
 @import url(./components/elements/TitleBlock/TitleBlock.css);
 @import url(./components/elements/Search/Search.css);
 
+@import url(./components/App/App.css);
 @import url(./components/Header/Header.css);
 @import url(./components/Hero/Hero.css);
 @import url(./components/Navigation/Navigation.css);
